Ménage, restructuration de folder, et quelques bug fix sur les nouveaux schémas inclus dans personnes et organisations
